Add generation applying error handling option and improve pipelines
An option is added to the settings to allow handling of generation result errors. The committed changes also improve the structure and readability of the code generation pipeline. Now, it's easier to follow what's happening during code generation process, improving maintainability significantly. The modification logic for source code during the pipeline process is abstracted as well.